Cleaned up the settings in the Activity settings project and localization of configuration in place
