+++ ActivitySettings/Configuration/Enums.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ActivitySettings.Configuration
+{
+    /// <summary>
+    /// Sample enumeration for drop down demonstration
+    /// </summary>
+    public enum EnumToDemo
+    {
+        EnumeratorValue1 = 0,
+        EnumeratorValue2 = 1,
+        EnumeratorValue3 = 2,
+        EnumeratorValue4 = 3,
+    }
+}
